fix(frontend): Add checks on draw control to prevent crashes
I was experiencing a number of crashes due to trying to use the
draw control when it is possible for it not to exist, so added
some checks to prevent this.

+++ client/src/satellites/satellite-search.component.js
@@ -64,7 +64,7 @@ const SatelliteSearch = ({ map, satellites, setVisiblePanel, setSelectedMoreInfo
     if (feature) {
       const featureArea = getGeometryAreaKmSquared(feature.geometry.coordinates[0]);
       const isTooLarge = featureArea > maximumAoiArea;
-      if (isTooLarge) {
+      if (drawControl && isTooLarge) {
         drawControl.setFeatureProperty(feature.id, 'error', 'true');
       }
     }
@@ -141,22 +141,29 @@ const SatelliteSearch = ({ map, satellites, setVisiblePanel, setSelectedMoreInfo
         const line = lineString(aoi);
         mapInstance.fitBounds(bbox(line), { padding: 275, offset: [100, 0] });
         const [drawCtrl] = getDraw();
-        drawCtrl.deleteAll();
-        const feature = {
-          type: 'Feature',
-          drawType: 'AOI',
-          properties: {
+        if (drawCtrl) {
+          drawCtrl.deleteAll();
+          const feature = {
+            type: 'Feature',
             drawType: 'AOI',
-            fillOpacity: 0.5,
-            fillColor: 'green'
-          },
-          geometry: {
-            type: 'Polygon',
-            coordinates: [aoi]
+            properties: {
+              drawType: 'AOI',
+              fillOpacity: 0.5,
+              fillColor: 'green'
+            },
+            geometry: {
+              type: 'Polygon',
+              coordinates: [aoi]
+            }
+          };
+          drawCtrl.add(feature);
+        }
+
+        return () => {
+          if (drawCtrl) {
+            drawCtrl.deleteAll();
           }
         };
-        drawCtrl.add(feature);
-        return () => drawCtrl.deleteAll();
       }
     },
     [currentSearchQuery]
